Extracted a part of IsContextEqualTo to a separate matcher
